Big refactor of PlanetAtom  components, models and logic; Split into smaller components;

+++ Assets/Scripts/Features/Planets/Components/PlanetAtomModel.cs
@@ -8,6 +8,6 @@ public class PlanetAtomModel
     public string Name;
     public int AtomicNumber;
     public int Stock;
-    public int HarvestRate;
+    public float HarvestRate;
     public int UpgradeLevel;
 }
